requirements.txt
# core
numpy>=1.24
pandas>=2.0
pyyaml>=6.0

# embeddings (for MiniLM via sentence-transformers)
sentence-transformers>=2.5.1
torch>=2.2  

# plotting
matplotlib>=3.7
